refactor(gh-936): extract shared dialog value helpers (cut new-code duplication)
safeStr/optFloat were copied from TedEditDialog into TurbulatorEditDialog;
move them to lib/dialogValues.ts and import, removing the duplicated block so
SonarCloud new_duplicated_lines_density drops below the 3% gate.
